+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv (development version)
 
+* Fixed an issue where `renv::dependencies()` could fail for chunks which
+  had a `label` which was not a length-one character vector. (#2144)
+
 * Fixed an issue where `file.rename()` could fail when unpacking a
   package during installation in certain Windows environments. (#2156)